Read ammo parameters from weapon files
Ammo parameters are read from static data in the BG ammoTable
(included by both Client and Game modules). This data is now loaded
into ammoTable from weapon files. Each .weap file should include an
ammo block as a parent of weaponDef - if not present then the static
data is used.

weaponDef
{
	client {
		...
	}

	ammo {
		maxammoPerSkill     400 300 200 150 150
		maxclipPerSkill     8   8   8   8   8
		uses                1
		reloadTime          1500
		fireDelayTime       100
		nextShotTime        300
		nextShotTime2       300
		maxHeat             0
		coolRate            0
		playerDamage        7
		aiDamage            5
		playerSplashRadius  0
		aiSplashRadius      0
		spread              400
		aimSpreadScaleAdd   35
		spreadScale         0.3
		weapRecoilDuration  50
		weapRecoilPitch     0.2 0.1
		weapRecoilYaw       0.0 0.0
		soundRange          700
		moveSpeed           0.95
	}
}

The ammotable_t maxammo and maxclip fields are specified for each skill
level using maxammoPerSkill and maxclipPerSkill. Five values are
expected in the order of easy, medium, hard, max and realism. These
skill-based values are set directly in the ammoTable when the skill
level is changed.