SP: Fix crash loading savegame
Clear cent->pe.legs.animation before calling CG_SetInitialSnapshot().

When using DLLs (the default) cent->pe.legs.animation points directly to
Game VM memory. The G_Alloc() memory pool address may be different after
the Game VM is reloaded and cause accessing the pointer to SIGSEGV.
(It may be the wrong contents anyway as the memory pool was reset.)

It crashed if the savegame is loaded while already on the same map, the
Game VM reloads with a different base address (ASLR), and the Cast AI
happens to be in the first snapshot (opposed to a later snapshot as
loading the savegame may not be complete in the first snapshot).

Fix by zturtleman
Synced up from iortcw.